Add files via upload
The main components of the ELK stack is explored by going through the installation and configuring process and linking them together to form an effective log analysis utility.
In this room, installing, configuring, and integrating different components of the ELK stack to form a complete log analysis utility. This document also covers creating alerts in ELK (Wazuh) and creating advanced KQL search queries to create complex search queries to improve the log investigation.